Move registry views to separate file
[noissue]
